// src/app/core/models/class.model.ts
export interface InstructorList {
  user_id: InstructorDetail[];
  instructor_id: String;
}
export interface InstructorDetail {
  name: String;
  last_name: String;
}

export interface LabDetail {
  name: string;
  _id: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
  id: string;
}

export interface LabListModel {
  docs: LabDetail[];
}

export interface CourseTitleModel extends Pagination {
  _id: string;
  id: string;
  title: string;
}
// ...
export interface DataSourceModel {
  start: string;
  end: string;
  instructor: string;
  lab: string;
}
export interface SessionModel {
  sessionNumber: number;
  sessionStartDate: string;
  sessionEndDate: string;
  sessionStartTime: string;
  sessionEndTime: string;
  instructorId: string;
  laboratoryId: string;
}